+++ doc/CHANGES.md
@@ -2,3 +2,7 @@
 
 ## Version 0.1.0 - Initial Release / Prototype
 This is the initial release of the Fate Language Prototype.  There's still quite a lot to do and probably more than a few bugs, but it's at a stage now where it's somewhat usable.
+
+## Version 0.2.0 - Static Single Assignment Form
+
+## Version 0.3.0 - Syntax / Grammar Stable